+++ SHEILa_runtime/src/Feeling/Feeling.h
@@ -37,47 +37,47 @@ public:
 
 	Feeling();
 	Feeling(std::string name, std::string desc,
-			uintmax_t ang_low, uintmax_t ang_hi,
-			uintmax_t dis_low, uintmax_t dis_hi,
-			uintmax_t sad_low, uintmax_t sad_hi,
-			uintmax_t sur_low, uintmax_t sur_hi,
-			uintmax_t fea_low, uintmax_t fea_hi,
-			uintmax_t tru_low, uintmax_t tru_hi,
-			uintmax_t joy_low, uintmax_t joy_hi,
-			uintmax_t ant_low, uintmax_t ant_hi,
-			double min_diff_ang_dis, double max_diff_ang_dis,
-			double min_diff_ang_sad, double max_diff_ang_sad,
-			double min_diff_ang_sur, double max_diff_ang_sur,
-			double min_diff_ang_fea, double max_diff_ang_fea,
-			double min_diff_ang_tru, double max_diff_ang_tru,
-			double min_diff_ang_joy, double max_diff_ang_joy,
-			double min_diff_ang_ant, double max_diff_ang_ant,
-			double min_diff_dis_sad, double max_diff_dis_sad,
-			double min_diff_dis_sur, double max_diff_dis_sur,
-			double min_diff_dis_fea, double max_diff_dis_fea,
-			double min_diff_dis_tru, double max_diff_dis_tru,
-			double min_diff_dis_joy, double max_diff_dis_joy,
-			double min_diff_dis_ant, double max_diff_dis_ant,
-			double min_diff_sad_sur, double max_diff_sad_sur,
-			double min_diff_sad_fea, double max_diff_sad_fea,
-			double min_diff_sad_tru, double max_diff_sad_tru,
-			double min_diff_sad_joy, double max_diff_sad_joy,
-			double min_diff_sad_ant, double max_diff_sad_ant,
-			double min_diff_sur_fea, double max_diff_sur_fea,
-			double min_diff_sur_tru, double max_diff_sur_tru,
-			double min_diff_sur_joy, double max_diff_sur_joy,
-			double min_diff_sur_ant, double max_diff_sur_ant,
-			double min_diff_fea_tru, double max_diff_fea_tru,
-			double min_diff_fea_joy, double max_diff_fea_joy,
-			double min_diff_fea_ant, double max_diff_fea_ant,
-			double min_diff_tru_joy, double max_diff_tru_joy,
-			double min_diff_tru_ant, double max_diff_tru_ant,
-			double min_diff_joy_ant, double max_diff_joy_ant);
+			long double ang_low, long double ang_hi,
+			long double dis_low, long double dis_hi,
+			long double sad_low, long double sad_hi,
+			long double sur_low, long double sur_hi,
+			long double fea_low, long double fea_hi,
+			long double tru_low, long double tru_hi,
+			long double joy_low, long double joy_hi,
+			long double ant_low, long double ant_hi,
+			long double min_diff_ang_dis, long double max_diff_ang_dis,
+			long double min_diff_ang_sad, long double max_diff_ang_sad,
+			long double min_diff_ang_sur, long double max_diff_ang_sur,
+			long double min_diff_ang_fea, long double max_diff_ang_fea,
+			long double min_diff_ang_tru, long double max_diff_ang_tru,
+			long double min_diff_ang_joy, long double max_diff_ang_joy,
+			long double min_diff_ang_ant, long double max_diff_ang_ant,
+			long double min_diff_dis_sad, long double max_diff_dis_sad,
+			long double min_diff_dis_sur, long double max_diff_dis_sur,
+			long double min_diff_dis_fea, long double max_diff_dis_fea,
+			long double min_diff_dis_tru, long double max_diff_dis_tru,
+			long double min_diff_dis_joy, long double max_diff_dis_joy,
+			long double min_diff_dis_ant, long double max_diff_dis_ant,
+			long double min_diff_sad_sur, long double max_diff_sad_sur,
+			long double min_diff_sad_fea, long double max_diff_sad_fea,
+			long double min_diff_sad_tru, long double max_diff_sad_tru,
+			long double min_diff_sad_joy, long double max_diff_sad_joy,
+			long double min_diff_sad_ant, long double max_diff_sad_ant,
+			long double min_diff_sur_fea, long double max_diff_sur_fea,
+			long double min_diff_sur_tru, long double max_diff_sur_tru,
+			long double min_diff_sur_joy, long double max_diff_sur_joy,
+			long double min_diff_sur_ant, long double max_diff_sur_ant,
+			long double min_diff_fea_tru, long double max_diff_fea_tru,
+			long double min_diff_fea_joy, long double max_diff_fea_joy,
+			long double min_diff_fea_ant, long double max_diff_fea_ant,
+			long double min_diff_tru_joy, long double max_diff_tru_joy,
+			long double min_diff_tru_ant, long double max_diff_tru_ant,
+			long double min_diff_joy_ant, long double max_diff_joy_ant);
 	virtual ~Feeling();
 	const std::string& getDescription() const;
 	const std::string& getName() const;
 
-	std::pair<double,double> getDiff(char a, char b);
+	std::pair<long double,long double> getDiff(char a, char b);
 
 	Range ang_range;
 	Range dis_range;
@@ -93,34 +93,34 @@ protected:
 	std::string name;
 	std::string description;
 
-	double min_diff_ang_dis; double max_diff_ang_dis;
-	double min_diff_ang_sad; double max_diff_ang_sad;
-	double min_diff_ang_sur; double max_diff_ang_sur;
-	double min_diff_ang_fea; double max_diff_ang_fea;
-	double min_diff_ang_tru; double max_diff_ang_tru;
-	double min_diff_ang_joy; double max_diff_ang_joy;
-	double min_diff_ang_ant; double max_diff_ang_ant;
-	double min_diff_dis_sad; double max_diff_dis_sad;
-	double min_diff_dis_sur; double max_diff_dis_sur;
-	double min_diff_dis_fea; double max_diff_dis_fea;
-	double min_diff_dis_tru; double max_diff_dis_tru;
-	double min_diff_dis_joy; double max_diff_dis_joy;
-	double min_diff_dis_ant; double max_diff_dis_ant;
-	double min_diff_sad_sur; double max_diff_sad_sur;
-	double min_diff_sad_fea; double max_diff_sad_fea;
-	double min_diff_sad_tru; double max_diff_sad_tru;
-	double min_diff_sad_joy; double max_diff_sad_joy;
-	double min_diff_sad_ant; double max_diff_sad_ant;
-	double min_diff_sur_fea; double max_diff_sur_fea;
-	double min_diff_sur_tru; double max_diff_sur_tru;
-	double min_diff_sur_joy; double max_diff_sur_joy;
-	double min_diff_sur_ant; double max_diff_sur_ant;
-	double min_diff_fea_tru; double max_diff_fea_tru;
-	double min_diff_fea_joy; double max_diff_fea_joy;
-	double min_diff_fea_ant; double max_diff_fea_ant;
-	double min_diff_tru_joy; double max_diff_tru_joy;
-	double min_diff_tru_ant; double max_diff_tru_ant;
-	double min_diff_joy_ant; double max_diff_joy_ant;
+	long double min_diff_ang_dis; long double max_diff_ang_dis;
+	long double min_diff_ang_sad; long double max_diff_ang_sad;
+	long double min_diff_ang_sur; long double max_diff_ang_sur;
+	long double min_diff_ang_fea; long double max_diff_ang_fea;
+	long double min_diff_ang_tru; long double max_diff_ang_tru;
+	long double min_diff_ang_joy; long double max_diff_ang_joy;
+	long double min_diff_ang_ant; long double max_diff_ang_ant;
+	long double min_diff_dis_sad; long double max_diff_dis_sad;
+	long double min_diff_dis_sur; long double max_diff_dis_sur;
+	long double min_diff_dis_fea; long double max_diff_dis_fea;
+	long double min_diff_dis_tru; long double max_diff_dis_tru;
+	long double min_diff_dis_joy; long double max_diff_dis_joy;
+	long double min_diff_dis_ant; long double max_diff_dis_ant;
+	long double min_diff_sad_sur; long double max_diff_sad_sur;
+	long double min_diff_sad_fea; long double max_diff_sad_fea;
+	long double min_diff_sad_tru; long double max_diff_sad_tru;
+	long double min_diff_sad_joy; long double max_diff_sad_joy;
+	long double min_diff_sad_ant; long double max_diff_sad_ant;
+	long double min_diff_sur_fea; long double max_diff_sur_fea;
+	long double min_diff_sur_tru; long double max_diff_sur_tru;
+	long double min_diff_sur_joy; long double max_diff_sur_joy;
+	long double min_diff_sur_ant; long double max_diff_sur_ant;
+	long double min_diff_fea_tru; long double max_diff_fea_tru;
+	long double min_diff_fea_joy; long double max_diff_fea_joy;
+	long double min_diff_fea_ant; long double max_diff_fea_ant;
+	long double min_diff_tru_joy; long double max_diff_tru_joy;
+	long double min_diff_tru_ant; long double max_diff_tru_ant;
+	long double min_diff_joy_ant; long double max_diff_joy_ant;
 
 };
